fix: delayed render fix moved to Switch component
