+++ src/screens/Profile/Appearance/index.tsx
@@ -16,18 +16,6 @@ export default function Appearance() {
   const style = useProfileHomeStyle();
   const {t} = useTranslation();
 
-  function useDelay(ms: number) {
-    const [gate, setGate] = useState(false);
-    useEffect(() => {
-      setTimeout(() => {
-        setGate(true);
-      }, ms);
-    });
-    return gate;
-  }
-
-  const delayRender = useDelay(100);
-
   return (
     <View style={style.container}>
       <FullScreenHeader
@@ -36,27 +24,25 @@ export default function Appearance() {
       />
 
       <ScrollView>
-        {delayRender && (
-          <Section withTopPadding withPadding>
+        <Section withTopPadding withPadding>
+          <ActionItem
+            mode="toggle"
+            text={t(AppearanceSettingsTexts.actions.usePhoneSettings)}
+            checked={!overrideColorScheme}
+            onPress={(checked) => overrideOSThemePreference(!checked)}
+          />
+
+          {overrideColorScheme && (
             <ActionItem
               mode="toggle"
-              text={t(AppearanceSettingsTexts.actions.usePhoneSettings)}
-              checked={!overrideColorScheme}
-              onPress={(checked) => overrideOSThemePreference(!checked)}
+              text={t(AppearanceSettingsTexts.actions.darkMode)}
+              checked={storedColorScheme === 'dark'}
+              onPress={(checked) =>
+                updateThemePreference(checked ? 'dark' : 'light')
+              }
             />
-
-            {overrideColorScheme && (
-              <ActionItem
-                mode="toggle"
-                text={t(AppearanceSettingsTexts.actions.darkMode)}
-                checked={storedColorScheme === 'dark'}
-                onPress={(checked) =>
-                  updateThemePreference(checked ? 'dark' : 'light')
-                }
-              />
-            )}
-          </Section>
-        )}
+          )}
+        </Section>
       </ScrollView>
     </View>
   );
